+++ src/azure-cli-core/HISTORY.rst
@@ -5,7 +5,7 @@ Release History
 2.0.27
 ++++++
 * auth: key on both subscription id and name on msi login
-* Minor fixes
+* Add events module in core for EVENT_INVOKER_PRE_CMD_TBL_TRUNCATE
 
 2.0.26
 ++++++